update the tty for gpg's ssh-agent in preexec
this ensures that the ssh-agent displays the password prompt on the
correct tty at the expense of making this call before every command. it
is not an ideal solution, but is the only way to ensure that the correct
tty is used, even when switching back and forth between various tmux
panes and windows.