add sshcontrol
i'm not entirely sure if this is sane to add to the repository, but
manually adding the keygrip for a gpg authentication key is necessary
in order to enable using the authentication key for ssh authentication.

yeah, i spent about half a day banging my head trying to figure out why
openssh wasn't proxying authentication requests to gpg-agent. for now,
this is the only key i manage, so it makes sense to store in sshcontrol.
if i end up needing to manage any standard ssh keys, these are added
automatically by ssh-add (and should differ per machine) so this patch
should be reverted and the file unmanaged.

makes it a little annoying, but i don't see another solution currently.