feat: remove git-crypt from the repository
`git-crypt` was used to encrypt entire files. It has several
limitations, including a lack of external secret storage (other than a
key file), the requirement to encrypt the entire contents of the file,
and poor support for key rotation, to name a few. While it is a
functional tool, I have moved on to using sops [1] to solve a number of
the aforementioned problems.

[1]: https://github.com/mozilla/sops