feat(polybar): set the default date display to be more reader-friendly
This commit inverts the `date` and `date-alt` properties of the `date`
module for polybar. Previously, the default date presentation model was
previously machine friendly, with output like:

2021-02-27 09:01:38-0700

The above has been refactored to be the _alternate_ presentation model.
The default presentation model is now more digestable for humans, with
output similar to:

Sat 27 Feb 2021 09:01