feat(zsh): remove custom function: chat
I've migrated to using weechat's inbuilt `sec.crypt.passphrase_command`
to invoke an external program and use its output to unlock the secure
store, so this function is unecessary.

An alias `chat` is created in order to support starting weechat, as this
is what I am used to typing on the command line.

Change-Id: Ib7dd9e6510d6c34bb7813af5dd3232f920eaec59
GitOrigin-RevId: 3bcffc7483c6040c626d4ac35369c1501241ad98