feat: migrate from `ytop` to `btm`
`ytop` [0] is no longer maintained; the author recommends using
`bottom` [1] as an alternative solution. after exploring `bottom`, it
was determined to be a more complete solution than `ytop`, as it
provides more detailed information and is able to be manipulated and
configured as desired. with that, it has become my preferred tool to
leave running in a window within the "monitoring" tmux session.

with `bottom`, we can simplify the polybar battery module to not display
the chargin/consumption rate or time left to full/empty.

[0]: https://github.com/cjbassi/ytop
[1]: https://github.com/ClementTsang/bottom

Change-Id: Idb6da69c090fcb0f71358b0ac0ba8b3d77bfa49f
GitOrigin-RevId: a9b010b93d0a1cc32b33d75bb1ccece7f4ae8c00