feat: keep git-related buffers after hiding
this change removes the behavior causing these buffers to be deleted
after hiding them (e.g. switching to a new buffer). this was desired
behavior in the past, however, usage has shown that it is often
preferred to keep the buffers around, as the message may be populated
using lines and text yanked from other buffers, or diffs may need to be
reviewed in order to write a helpful commit message.

Change-Id: Ic961ca90a67ff812512a6f3ccf576e8f127047f4
GitOrigin-RevId: 7edc47773a08fb772b3513c85b1ebe9b1b10e41e