refactor: use the `submit` alias instead of `push --force`
This change replaces invocations of `git push --force` with `git
submit`, which is an alias that pushes to the correct desitination ref
for creating changes in a Gerrit-styled workflow. All aliases were
changed except for `git-pf`, which was ignored as it is used as
shorthand when manually patching repositories and is not used in the
typical submission process.

By default, the `submit` alias pushes to `refs/for/trunk`, but this can
be reconfigured in local git configuration files when a different
destination is desired.

Change-Id: I34a279ad4f5db494f0c532e69973766fd98f6f95
GitOrigin-RevId: 8bb9fc751fce03e45c664900166f125cf0e828c5